# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAchat
  description: WASAchat backend REST API
  version: 0.0.1
servers:
- url: http://localhost:3000
components:
  schemas:
    User:
      title: User
      description: "app user"
      type: object
      properties:
        id: 
          description: Unique identifier of a user in the app
          type: integer
          minimum: 0
          readOnly: true
        username: 
          description: Name choose by the user that can represent him in the app
          type: string
          pattern: "^[a-zA-Z0-9_.]+$"
          minLength: 3
          maxLength: 16
        profileImage:
          description: Image that represent the user
          format: binary
          type: string
          minLength: 0
          maxLength: 10000000
        realName:
          description: Real name of the user
          type: string
          pattern: "^[a-zA-Z]+$"
          minLength: 3
          maxLength: 16
        realSurname:
          description: Real surname of the user
          type: string
          pattern: "^[a-zA-Z]+$"
          minLength: 3
          maxLength: 16
    Message:
      title: Message
      description: Text or image data sent from a user to another
      type: object
      properties:
        id: 
          description: Unique identifier of a message in the chat
          type: integer 
          readOnly: true
        messageType:
          description: Identifier of the type of the message
          type: string
          pattern: "^[a-zA-Z]+$"
          enum: ["image", "text"]
          minLength: 4
          maxLength: 5
        content: 
          description: Content of the message
          type: string
          pattern: ".*"
          minLength: 1
          maxLength: 8000
        previewContent: 
          description: content
          type: string
          pattern: ".*"
          minLength: 1
          maxLength: 32
        sender:
          description: username of the user that have sent the message
          type: string
          pattern: "^[a-zA-Z0-9_.]+$"
          minLength: 3
          maxLength: 16
        timestamp:
          description: Unix timestamp of the time when the user sent the message
          type: integer
        status: 
          description: Show if the user have received the message or not
          type: string
          pattern: "^[a-zA-Z]+$"
          enum: ["sent", "received", "not_sent"]
          minLength: 4
          maxLength: 8
        replyTo:
          description: id of the message who this message reply
          type: integer
          minimum: 0
    Chat:
      title: Chat
      description: Conversations between 2 or more users
      type: object
      properties:
        id: 
          description: Unique identifier of the chat from an user
          type: integer
          minimum: 0
        messageList:
          description: Message list in the chat
          type: array
          items:
            $ref: '#/components/schemas/Message'
          minItems: 1
          maxItems: 1000000
        partecipantsId:
          description: list of users ID who take part of the chat
          type: array
          items:
            description: user id
            type: integer
            minimum: 0
          minItems: 2
          maxItems: 100
        photo: 
          description: Photo of the chat if is a group
          type: string
          format: binary
          minLength: 0
          maxLength: 10000000
  securitySchemes:
    BearerAuth: 
      type: http 
      scheme: bearer 
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /users/{id}/photo:
    parameters:
        - name: id
          in: path
          description: The ID of the user who you want to change or add photo
          schema:
            description: An integer representing the ID of the user who you want to change or add photo
            type: integer
            minimum: 0
          required: true
    post: 
      tags: ["User Management"]
      summary: Add or Change Photo
      description: Add or Update of the Photo of the user for this service
      operationId: setMyPhoto
      requestBody:
        description: User new Photo over the last one
        content:
          multipart/form-data:
            schema:
              description: User new Photo over the last one
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: The photo file to upload (image)
                  minLength: 0
                  maxLength: 10000000
        required: true
      responses:
        "200":
          description: photo updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
        "401":
          description: User have to Auth itself to set a photo
        "404": 
          description: User not found
  /users/{id}/conversations/{conversationId}/users:
    parameters:
      - name: id
        in: path
        description: The ID of the user
        schema:
          description: An int representing the ID of the user 
          type: integer
          minimum: 0
        required: true
      - name: conversationId
        in: path
        description: The ID of the group chat (Chat) to add the user to
        schema:
          description: ID of the chat
          type: integer
        required: true
    post:
      tags: ["Groupchat Management"]
      summary: Add a user to a group chat (Chat)
      description: Adds a user to an existing group chat, which is a conversation with isGroup set to true
      operationId: addToGroup
      requestBody:
        description: The user details to add to the group chat
        content:
          application/json: 
            schema:
              description: Object of the user to add or remove from the group chat
              type: object
              properties:
                userId: 
                  type: integer
                  description: The ID of the user to add to the group chat
        required: true
      responses:
        '200':
          description: User successfully added to the group chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        "400":
          description: Invalid input (e.g. invalid user ID)
        "401":
          description: User must authenticate to add someone to a group
        "403":
          description: User does not have permission to add users to this group
        "404":
          description: group chat or user not found
    delete:
      tags: ["Groupchat Management"]
      summary: Leave a group chat (Chat)
      description: Allows a user to leave an existing group chat (Chat).
      operationId: leaveGroup
      responses:
        '204':
          description: User successfully left the group chat
        "400":
          description: Invalid input (e.g. user not part of the conversation)
        "401":
          description: User must authenticate to leave the group
        #"403": non so se metterla
        #  description: User is not authorized to leave the group
        "404":
          description: Conversation not found or user is not a participant
...
  /users/{id}/conversations/{conversationId}/photo:
    post: 
      tags: ["Groupchat Management"]
      summary: Change GroupChat (Chat) Photo
      description: Update of the Photo of the GroupChat (Chat) selected
      operationId: setGroupPhoto
      parameters:
        - name: id
          in: path
          description: The ID of the user 
          schema:
            description: An integer representing the ID of the user
            type: integer
            minimum: 0
          required: true
        - name: conversationId
          in: path
          description: The ID of the group chat (Chat) the user is leaving
          schema:
            description: The ID of the group chat (Chat)
            type: integer
          required: true
      requestBody:
        description: Group new Photo over the last one
        content:
          multipart/form-data: 
            schema:
              description: New Photo 
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: The photo file to upload (image)
                  minLength: 0
                  maxLength: 10000000
        required: true
      responses:
        "200":
          description: photo updated successfully
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Chat"
        "400":
          description: Invalid input
        "401":
          description: User have to Auth itself to set a photo
        "404": 
          description: User not found
...